code put in 1 threath

+++ sw/airborne/modules/computer_vision/practical_module.h
@@ -72,7 +72,7 @@ float heading_goal_f;
 float heading_goal_ref;
 float r_dot_new_cos;
 float r_dot_new_sin;
-extern uint16_t matrix_treshold;
+extern float matrix_treshold;
 extern uint16_t matrix_sum_treshold;
 extern float ref_pitch_angle;
 extern int size_matrix[3];
